# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для мессенджера.",
        "title": "Messenger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/messages/voice": {
            "get": {
                "description": "Возвращает список голосовых сообщений между двумя пользователями",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "voice"
                ],
                "summary": "Получение голосовых сообщений",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID отправителя",
                        "name": "sender_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID получателя",
                        "name": "receiver_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.VoiceMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Отправляет голосовое сообщение от одного пользователя к другому",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "voice"
                ],
                "summary": "Отправка голосового сообщения",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID отправителя",
                        "name": "sender_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID получателя",
                        "name": "receiver_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Аудиофайл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "config.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
        "config.SimpleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Операция выполнена успешно"
                }
            }
        },
...
        "config.VoiceMessage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}